commit29

+++ src/main/java/com/doctor/appointment/service/impl/AppointmentImpl.java
@@ -1,6 +1,8 @@
 package com.doctor.appointment.service.impl;
 
 import com.doctor.appointment.dto.appointment.*;
+import com.doctor.appointment.exception.AppointmentAlreadyBooked;
+import com.doctor.appointment.exception.CannotDeleteAppointmentException;
 import com.doctor.appointment.mapper.MapperAppointment;
 import com.doctor.appointment.model.Appointment;
 import com.doctor.appointment.repository.AppointmentRepository;
@@ -16,18 +18,18 @@ import java.util.stream.Collectors;
 public class AppointmentImpl implements AppointmentService {
 
     private AppointmentRepository appointmentRepository;
-    @Override
-    public CreateResponseAppointmentDto create(CreateRequestAppointmentDto appointmentDto) {
-        Appointment createdAppointment = MapperAppointment.toEntity(appointmentDto);
-        Appointment appointmentResponse = appointmentRepository.save(createdAppointment);
-        return MapperAppointment.toDto(appointmentResponse);
-    }
 
     @Override
     public Appointment readById(long id) {
         return appointmentRepository.findById(id).orElseThrow(
                 () -> new EntityNotFoundException("Appointment " + id + " not found!!!")
         );
+    }
+    @Override
+    public CreateResponseAppointmentDto create(CreateRequestAppointmentDto appointmentDto) {
+        Appointment createdAppointment = MapperAppointment.toEntity(appointmentDto);
+        Appointment appointmentResponse = appointmentRepository.save(createdAppointment);
+        return MapperAppointment.toDto(appointmentResponse);
     }
 
    @Override
@@ -41,6 +43,9 @@ public class AppointmentImpl implements AppointmentService {
     @Override
     public void delete(long id) {
         Appointment appointment = readById(id);
+        if (appointment.getPatient() != null) {
+            throw new CannotDeleteAppointmentException("Appointment with existing patient cannot be deleted");
+        }
         appointmentRepository.delete(appointment);
     }
 
@@ -54,6 +59,9 @@ public class AppointmentImpl implements AppointmentService {
     @Override
     public ToMakeResponseAppointmentDto toMake(ToMakeRequestAppointmentDto dto, long id) {
         Appointment found = readById(id);
+        if (found.getPatient() != null) {
+            throw new AppointmentAlreadyBooked("Appointment already booked");
+        }
         Appointment appointment = MapperAppointment.toMakeToEntity(dto, found);
         Appointment appointmentResponse = appointmentRepository.save(appointment);
         return MapperAppointment.toMakeToDto(appointmentResponse);
